+++ EveryEuro/src/services/service.py
@@ -5,21 +5,22 @@ from tkinter.messagebox import *
 ALL_MONTHS = ["", "JANUARY", "FEBRUARY", "MARCH", "APRIL", "MAY", "JUNE", "JULY", "AUGUST",
                 "SEPTEMBER", "OCTOBER", "NOVEMBER", "DECEMBER"]
 
-def calculate_left_to_budget(income, bills, spending, debt):
-    left_to_budget = income - bills - spending - debt
+def calculate_left_to_budget(income, rent, bills, spending, debt_service, saving):
+    left_to_budget = income - rent - bills - spending - debt_service
     print(f"left_to_budget is: {left_to_budget}")
     return left_to_budget
 
+def bar_item_notdone():
+    showerror('Not implemented', 'Functionality not yet available, but coming soon')
+
+# combine the two functions below
 def get_month_name(month_number):
     all_months = ["", "JANUARY", "FEBRUARY", "MARCH", "APRIL", "MAY", "JUNE", "JULY", "AUGUST",
                     "SEPTEMBER", "OCTOBER", "NOVEMBER", "DECEMBER"]
     return all_months[month_number]
 
-def bar_item_notdone():
-    showerror('Not implemented', 'Functionality not yet available, but coming soon')
-
 def get_month_number_and_name(chosen_month):
-    month_name = chosen_month #self._chosen_month.get()
+    month_name = chosen_month
     dict = {"JANUARY": 1, "FEBRUARY": 2, "MARCH": 3, "APRIL": 4,
             "MAY": 5,"JUNE": 6, "JULY": 7, "AUGUST": 8,
             "SEPTEMBER": 9, "OCTOBER": 10, "NOVEMBER": 11, "DECEMBER": 12}